+++ app/index.tsx
@@ -26,6 +26,8 @@ import { useToast } from "@/components/ui/use-toast"
 import { SiteHeader } from "@/components/site-header"
 import { ThemeProvider } from "@/components/theme-provider"
 
+import { NFTProvider } from "./nft/useNFT"
+
 interface RootLayoutProps {
   children: React.ReactNode
 }
@@ -418,11 +420,13 @@ export default function Index({ children }: RootLayoutProps) {
         }}
       >
         <ThemeProvider attribute="class" defaultTheme="light" enableSystem>
-          <div className="relative flex min-h-screen flex-col">
-            <SiteHeader />
-            <section className="container px-4 mt-4">{children}</section>
-          </div>
-          <Toaster />
+          <NFTProvider>
+            <div className="relative flex min-h-screen flex-col">
+              <SiteHeader />
+              <section className="container px-4 mt-4">{children}</section>
+            </div>
+            <Toaster />
+          </NFTProvider>
         </ThemeProvider>
       </AppContext.Provider>
     </>
